+++ packages/authhero/src/routes/auth-api/index.ts
@@ -2,3 +2,4 @@ export * from "./well-known";
 export * from "./token";
 export * from "./logout";
 export * from "./userinfo";
+export * from "./dbconnections";