+++ packages/drizzle/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @authhero/drizzle
 
+## 0.1.70
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.33.0
+
+## 0.1.69
+
+### Patch Changes
+
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.32.1
+
 ## 0.1.68
 
 ### Patch Changes