+++ packages/kysely/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @authhero/kysely-adapter
 
+## 0.25.3
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.33.0
+
+## 0.25.2
+
+### Patch Changes
+
+- update all build packages
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.32.1
+
 ## 0.25.1
 
 ### Patch Changes